refactor: HttpResponse 응답 리팩토링
