+++ tomcat/src/main/java/nextstep/jwp/controller/register/RegisterController.java
@@ -1,13 +1,15 @@
 package nextstep.jwp.controller.register;
 
+import javassist.NotFoundException;
 import nextstep.jwp.controller.base.AbstractController;
 import nextstep.jwp.service.UserService;
 import org.apache.coyote.http11.request.HttpRequest;
 import org.apache.coyote.http11.response.HttpResponse;
 import org.apache.coyote.http11.response.header.ContentType;
-import org.apache.coyote.http11.response.header.Header;
 import org.apache.coyote.http11.response.header.Status;
 
+import java.io.IOException;
+import java.net.URISyntaxException;
 import java.util.Map;
 
 public class RegisterController extends AbstractController {
@@ -29,10 +31,10 @@ public class RegisterController extends AbstractController {
     }
 
     @Override
-    protected HttpResponse doPost(final HttpRequest httpRequest) {
+    protected HttpResponse doPost(final HttpRequest httpRequest) throws NotFoundException, IOException, URISyntaxException {
         Map<String, String> params = httpRequest.getBody().getParams();
         userService.register(params.get("account"), params.get("password"), params.get("email"));
 
-        return HttpResponse.found(ContentType.HTML, Status.FOUND, Map.of(Header.LOCATION, "/index.html"), "");
+        return HttpResponse.withResource(Status.FOUND, "/index.html");
     }
 }
