+++ tomcat/src/test/java/org/apache/coyote/http11/request/header/ParamsTest.java
@@ -1,5 +1,6 @@
 package org.apache.coyote.http11.request.header;
 
+import org.apache.coyote.http11.request.body.Params;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
